+++ src/EmbDB/EmbDB/BPInnerNodeSetROSimpleCompressor.h
@@ -51,7 +51,7 @@ namespace embDB
 			return true;
 		}
 	private:
-		size_t m_nSize;
+		uint32 m_nSize;
 	};
 }
